BASELINE: Update Chromium to 51.0.2704.41
Also adds in all smaller components by reversing logic for exclusion.

+++ chromium/ipc/unix_domain_socket_util.cc
@@ -5,10 +5,7 @@
 #include "ipc/unix_domain_socket_util.h"
 
 #include <errno.h>
-#include <fcntl.h>
-#include <stddef.h>
 #include <sys/socket.h>
-#include <sys/stat.h>
 #include <sys/un.h>
 #include <unistd.h>
 
@@ -27,37 +24,25 @@ static_assert(sizeof(((sockaddr_un*)0)->sun_path) >= kMaxSocketNameLength,
 
 namespace {
 
-// Returns fd (>= 0) on success, -1 on failure. If successful, fills in
-// |unix_addr| with the appropriate data for the socket, and sets
-// |unix_addr_len| to the length of the data therein.
-int MakeUnixAddrForPath(const std::string& socket_name,
-                        struct sockaddr_un* unix_addr,
-                        size_t* unix_addr_len) {
+// This function fills in |unix_addr| with the appropriate data for the socket,
+// and sets |unix_addr_len| to the length of the data therein.
+// Returns true on success, or false on failure (typically because |socket_name|
+// violated the naming rules).
+bool MakeUnixAddrForPath(const std::string& socket_name,
+                         struct sockaddr_un* unix_addr,
+                         size_t* unix_addr_len) {
   DCHECK(unix_addr);
   DCHECK(unix_addr_len);
 
   if (socket_name.length() == 0) {
     LOG(ERROR) << "Empty socket name provided for unix socket address.";
-    return -1;
+    return false;
   }
   // We reject socket_name.length() == kMaxSocketNameLength to make room for
   // the NUL terminator at the end of the string.
   if (socket_name.length() >= kMaxSocketNameLength) {
     LOG(ERROR) << "Socket name too long: " << socket_name;
-    return -1;
-  }
-
-  // Create socket.
-  base::ScopedFD fd(socket(AF_UNIX, SOCK_STREAM, 0));
-  if (!fd.is_valid()) {
-    PLOG(ERROR) << "socket";
-    return -1;
-  }
-
-  // Make socket non-blocking
-  if (HANDLE_EINTR(fcntl(fd.get(), F_SETFL, O_NONBLOCK)) < 0) {
-    PLOG(ERROR) << "fcntl(O_NONBLOCK)";
-    return -1;
+    return false;
   }
 
   // Create unix_addr structure.
@@ -66,7 +51,36 @@ int MakeUnixAddrForPath(const std::string& socket_name,
   strncpy(unix_addr->sun_path, socket_name.c_str(), kMaxSocketNameLength);
   *unix_addr_len =
       offsetof(struct sockaddr_un, sun_path) + socket_name.length();
-  return fd.release();
+  return true;
+}
+
+// This functions creates a unix domain socket, and set it as non-blocking.
+// If successful, |out_fd| will be set to the new file descriptor, and the
+// function will return true. Otherwise returns false.
+bool CreateUnixDomainSocket(base::ScopedFD* out_fd) {
+  DCHECK(out_fd);
+
+  // Create the unix domain socket.
+  base::ScopedFD fd(socket(AF_UNIX, SOCK_STREAM, 0));
+  if (!fd.is_valid()) {
+    PLOG(ERROR) << "Failed to create AF_UNIX socket.";
+    return false;
+  }
+
+  // Now set it as non-blocking.
+  if (!base::SetNonBlocking(fd.get())) {
+    PLOG(ERROR) << "base::SetNonBlocking() failed " << fd.get();
+    return false;
+  }
+
+  fd.swap(*out_fd);
+
+  return true;
+}
+
+bool IsRecoverableError() {
+  return errno == ECONNABORTED || errno == EMFILE || errno == ENFILE ||
+         errno == ENOMEM || errno == ENOBUFS;
 }
 
 }  // namespace
@@ -75,21 +89,14 @@ bool CreateServerUnixDomainSocket(const base::FilePath& socket_path,
                                   int* server_listen_fd) {
   DCHECK(server_listen_fd);
 
-  std::string socket_name = socket_path.value();
-  base::FilePath socket_dir = socket_path.DirName();
-
-  struct sockaddr_un unix_addr;
-  size_t unix_addr_len;
-  base::ScopedFD fd(
-      MakeUnixAddrForPath(socket_name, &unix_addr, &unix_addr_len));
-  if (!fd.is_valid())
-    return false;
-
   // Make sure the path we need exists.
+  base::FilePath socket_dir = socket_path.DirName();
   if (!base::CreateDirectory(socket_dir)) {
     LOG(ERROR) << "Couldn't create directory: " << socket_dir.value();
     return false;
   }
+
+  const std::string socket_name = socket_path.value();
 
   // Delete any old FS instances.
   if (unlink(socket_name.c_str()) < 0 && errno != ENOENT) {
@@ -97,16 +104,25 @@ bool CreateServerUnixDomainSocket(const base::FilePath& socket_path,
     return false;
   }
 
+  struct sockaddr_un unix_addr;
+  size_t unix_addr_len;
+  if (!MakeUnixAddrForPath(socket_name, &unix_addr, &unix_addr_len))
+    return false;
+
+  base::ScopedFD fd;
+  if (!CreateUnixDomainSocket(&fd))
+    return false;
+
   // Bind the socket.
   if (bind(fd.get(), reinterpret_cast<const sockaddr*>(&unix_addr),
            unix_addr_len) < 0) {
-    PLOG(ERROR) << "bind " << socket_path.value();
+    PLOG(ERROR) << "bind " << socket_name;
     return false;
   }
 
   // Start listening on the socket.
   if (listen(fd.get(), SOMAXCONN) < 0) {
-    PLOG(ERROR) << "listen " << socket_path.value();
+    PLOG(ERROR) << "listen " << socket_name;
     unlink(socket_name.c_str());
     return false;
   }
@@ -119,14 +135,13 @@ bool CreateClientUnixDomainSocket(const base::FilePath& socket_path,
                                   int* client_socket) {
   DCHECK(client_socket);
 
-  std::string socket_name = socket_path.value();
-  base::FilePath socket_dir = socket_path.DirName();
-
   struct sockaddr_un unix_addr;
   size_t unix_addr_len;
-  base::ScopedFD fd(
-      MakeUnixAddrForPath(socket_name, &unix_addr, &unix_addr_len));
-  if (!fd.is_valid())
+  if (!MakeUnixAddrForPath(socket_path.value(), &unix_addr, &unix_addr_len))
+    return false;
+
+  base::ScopedFD fd;
+  if (!CreateUnixDomainSocket(&fd))
     return false;
 
   if (HANDLE_EINTR(connect(fd.get(), reinterpret_cast<sockaddr*>(&unix_addr),
@@ -177,20 +192,16 @@ bool IsPeerAuthorized(int peer_fd) {
   return true;
 }
 
-bool IsRecoverableError(int err) {
-  return errno == ECONNABORTED || errno == EMFILE || errno == ENFILE ||
-         errno == ENOMEM || errno == ENOBUFS;
-}
-
 bool ServerAcceptConnection(int server_listen_fd, int* server_socket) {
   DCHECK(server_socket);
   *server_socket = -1;
 
   base::ScopedFD accept_fd(HANDLE_EINTR(accept(server_listen_fd, NULL, 0)));
   if (!accept_fd.is_valid())
-    return IsRecoverableError(errno);
-  if (HANDLE_EINTR(fcntl(accept_fd.get(), F_SETFL, O_NONBLOCK)) < 0) {
-    PLOG(ERROR) << "fcntl(O_NONBLOCK) " << accept_fd.get();
+    return IsRecoverableError();
+
+  if (!base::SetNonBlocking(accept_fd.get())) {
+    PLOG(ERROR) << "base::SetNonBlocking() failed " << accept_fd.get();
     // It's safe to keep listening on |server_listen_fd| even if the attempt to
     // set O_NONBLOCK failed on the client fd.
     return true;
